+++ RS-School-React/package.json
@@ -15,7 +15,8 @@
     "axios-cache-interceptor": "^1.5.3",
     "pokenode-ts": "^1.20.0",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.24.1"
   },
   "devDependencies": {
     "@types/react": "^18.3.3",
