+++ RS-School-React/src/components/Results/Card/Card.css
