+++ RS-School-React/src/components/Results/Results.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable no-mixed-spaces-and-tabs */
 import { ItemInterface } from "../../interfaces/interfaces";
-import Result from "./Result/Result";
+import Card from "./Card/Card";
 import "./Results.css"
 
 interface Props {
@@ -9,16 +9,17 @@ interface Props {
 }
 
 const Results = (props: Props) => {
-	
 
     return (
-      <div className="results">
-        {props.items.map((item, id) => (
-          <div key={id}>
-            <Result name={item.name} url={item.url} />
-          </div>
-        ))}
-      </div>
+		<div className="bottom">
+			<div className="results">
+			{props.items.map((item, id) => (
+				<div key={id}>
+					<Card name={item.name} url={item.url} />
+				</div>
+			))}
+			</div>
+		</div>
     );
   }
