+++ RS-School-React/src/components/common/Page404/Page404.css
@@ -1,6 +1,5 @@
-.error-page {
-	font-size: 60px;
-	color: #265470;
-	height: 100%;
-	text-align: center;
+.error-page img {
+	display: block;
+	margin: 100px auto;
+	height: 300px;
 }
